refractor: modification de la page App pour la remettre au goût du jour, en créant une page spécifique dédiée aux status + vidage de styles.css pour utilisation de tailwind + ajout des composants Navbar et Footer, vierges pour l'instant
