+++ .gitignore
@@ -1,0 +1,3 @@
+node_modules/
+frontend/node_modules/
+backend/node_modules/
