+++ frontend/package.json
@@ -11,11 +11,15 @@
     "react-dom": "^18.3.1"
   },
   "devDependencies": {
+    "@tailwindcss/postcss": "^4.1.13",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.6.3",
+    "postcss": "^8.5.6",
+    "postcss-loader": "^8.1.1",
     "style-loader": "^4.0.0",
+    "tailwindcss": "^4.1.13",
     "ts-loader": "^9.5.2",
     "typescript": "^5.8.3",
     "webpack": "^5.99.9",
