+++ frontend/src/components/Footer.tsx
@@ -1,0 +1,6 @@
+export default function Footer() {
+  return (
+    <footer>
+    </footer>
+  );
+}
