+++ frontend/src/components/Navbar.tsx
@@ -1,0 +1,6 @@
+export default function Navbar() {
+  return (
+    <nav>
+    </nav>
+  );
+}
