+++ frontend/src/styles.css
@@ -1,3 +1,5 @@
+@import "tailwindcss";
+
 :root {
   font-family: "Space Grotesk", "Segoe UI", sans-serif;
   color: #f6f1e8;
@@ -12,91 +14,9 @@
 
 body {
   margin: 0;
-  min-height: 100vh;
 }
 
 a {
   color: inherit;
-}
-
-.page {
-  min-height: 100vh;
-  display: grid;
-  place-items: center;
-  padding: 24px;
-}
-
-.card {
-  width: min(760px, 100%);
-  border-radius: 24px;
-  padding: 32px;
-  backdrop-filter: blur(12px);
-  background: rgba(9, 19, 27, 0.72);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  box-shadow: 0 20px 60px rgba(0, 0, 0, 0.25);
-}
-
-.card--success {
-  box-shadow: 0 20px 60px rgba(46, 196, 182, 0.2);
-}
-
-.card--warning {
-  box-shadow: 0 20px 60px rgba(231, 111, 81, 0.25);
-}
-
-.card--error {
-  box-shadow: 0 20px 60px rgba(214, 40, 40, 0.25);
-}
-
-.eyebrow {
-  margin: 0 0 12px;
-  text-transform: uppercase;
-  letter-spacing: 0.18em;
-  font-size: 0.75rem;
-  color: #f4a261;
-}
-
-h1,
-h2,
-p {
-  margin: 0;
-}
-
-h1 {
-  font-size: clamp(2rem, 5vw, 3.4rem);
-  line-height: 1;
-  margin-bottom: 16px;
-}
-
-h2 {
-  margin-bottom: 12px;
-  color: #f4a261;
-}
-
-p {
-  max-width: 52ch;
-  line-height: 1.6;
-}
-
-.actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 12px;
-  margin: 24px 0;
-}
-
-.actions a {
   text-decoration: none;
-  padding: 12px 18px;
-  border-radius: 999px;
-  background: rgba(244, 162, 97, 0.16);
-  border: 1px solid rgba(244, 162, 97, 0.4);
-}
-
-pre {
-  margin: 0;
-  padding: 18px;
-  border-radius: 18px;
-  background: rgba(0, 0, 0, 0.28);
-  overflow: auto;
 }
